Assert at least one label for training
Added an assert statement to ensure that there is at least one label and roi
for the minibatch. If this is not the case, the training starts however results
in a floating point error which doesn't give the clear cause. This situation
can result from having images which are entirely negatives but not having
TRAIN.BG_THRESH_LO set to zero.